+++ src/components/card/CardContainer.jsx
@@ -1,4 +1,4 @@
-import Card from "@/components/Card"
+import Card from "@/components/card/Card"
 
 
 const CardContainer = ({ productsData }) => {
